+++ metadata.go
@@ -2,6 +2,6 @@ package slack
 
 // SlackMetadata https://api.slack.com/reference/metadata
 type SlackMetadata struct {
-	EventType    string                 `json:"event_type"`
-	EventPayload map[string]interface{} `json:"event_payload"`
+	EventType    string         `json:"event_type"`
+	EventPayload map[string]any `json:"event_payload"`
 }