+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **`MsgOptionPostMessageParameters` now passes `MetaData`** - Previously, metadata
+  was silently dropped when using `PostMessageParameters`. ([#1343])
+
 ## [0.18.0-rc1] - 2026-01-26
 
 ### Added